more audiotype clean up
Former-commit-id: 54fcbb91da8e1351d995eb62d54bd2f3c605b82f
Former-commit-id: 01dfa14b5019464b8e14e632ecc4cd834f0aa7f7 [formerly bd3ef773b9d506064968be1dd3ffdc0673279495]
Former-commit-id: c50adda0a7993edca54dfb2f72d9cb73b91d7024

+++ src/mitll/langtest/server/database/user/UserManagement.java
@@ -274,11 +274,11 @@ public class UserManagement {
    * Adds some sugar -- sets the answers and rate per user, and joins with dli experience data
    *
    * @return
-   * @see mitll.langtest.server.LangTestDatabaseImpl#getUsers()
+   * @see mitll.langtest.server.database.DatabaseImpl#getUsers
    * @see #usersToXLSX(OutputStream, String)
    */
   public List<User> getUsers() {
-    Map<Long, Float> userToRate = resultDAO.getSessions().getUserToRate();
+    Map<Integer, Float> userToRate = resultDAO.getSessions().getUserToRate();
     List<User> users = null;
     try {
       Pair idToCount = populateUserToNumAnswers();
